+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - 修复悬浮预览窗口拖拽失效问题 | Fixed the issue that dragging and dropping feature on the popover preview window failed.
 - 抽象并封装拖动功能 | Abstract and encapsulate the dragfeature.
 - 悬浮预览窗口支持使用鼠标拖动调整尺寸 | Popover preview window supports using the mouse to drag to adjust the size.
+- 支持自定义预览超链接目标的小窗口样式 | Supports custom small window styles for previewing hyperlinks target.
+- 支持设置默认文档访问权限 | Supports setting the default document access permission.
 
 ## v0.1.0 / 2022-07-25